add table turnos

+++ src/main/java/barberapp/repository/BarberoRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BarberoRepository extends JpaRepository<Barbero, Integer> {
+public interface BarberoRepository extends JpaRepository<Barbero, Long> {
     boolean existsByEmail(String email);
 }
